TST: Rework pressure trace reconstruction tests (#1013)
* TST: Rework pressure trace reconstruction tests

* STY: black and isort

* MAINT: Ignores (F401) imported but unused on files __init__.py

---------
