+++ setup.cfg
@@ -26,7 +26,9 @@ ignore =
 exclude = 
   src/porepy/__init__.py,
   src/porepy/numerics/ad/__init__.py,
-
+per-file-ignores =
+  # F401: imported but unused on files __init__.py
+  __init__.py:F401
   
 [mypy]
 warn_unused_configs = True
