+++ tests/numerics/fv/test_mpfa.py
@@ -432,3 +432,248 @@ def test_hydrostatic_pressure_2d(grid_class):
 def test_mpfa_gravity_common_with_tpfa(test_method):
     """See test_utils.common_xpfa_tests.py for the original tests."""
     test_method("mpfa")
+
+
+discr_instance = pp.Mpfa("flow")
+
+
+class TestMpfaBoundaryPressure(xpfa_tests.XpfaBoundaryPressureTests):
+    """Tests for the boundary pressure computation in MPFA.
+    Provides access to the fixture discr_instance and adds two tests for simplices.
+    Otherwise identical to the tests in test_utils.common_xpfa_tests.py and used in
+    test_tpfa.py.
+
+    """
+
+    @property
+    def discr_instance(self):
+        """Return a tpfa instance."""
+        return discr_instance
+
+    def test_linear_flow_simplex_grid(self):
+        mesh_size = {"mesh_size_frac": 0.3, "mesh_size_bound": 0.3}
+        network = pp.create_fracture_network(
+            None, pp.grids.standard_grids.utils.unit_domain(2)
+        )
+        mdg = network.mesh(mesh_size)
+        g = mdg.subdomains(dim=2)[0]
+        # Flow from right to left
+        bf = g.get_boundary_faces()
+        bc_type = np.asarray(bf.size * ["neu"])
+
+        xf = g.face_centers[:, bf]
+        xleft = np.where(xf[0] < 1e-3 + xf[0].min())[0]
+        xright = np.where(xf[0] > xf[0].max() - 1e-3)[0]
+        bc_type[xleft] = "dir"
+
+        bound = pp.BoundaryCondition(g, bf, bc_type)
+
+        bc_val = np.zeros(g.num_faces)
+        bc_val[bf[xright]] = 1 * g.face_areas[bf[xright]]
+
+        data = self.make_dictionary(g, bound, bc_val)
+
+        p = self.pressure(g, data)
+
+        bound_p = self.boundary_pressure(p, bc_val, data)
+        assert np.allclose(bound_p[bf], -g.face_centers[0, bf])
+
+    def test_structured_simplex_linear_flow(self):
+        g = self.simplex_grid()
+        bc_val = np.zeros(g.num_faces)
+        # Flow from rght to left
+        bf = g.get_boundary_faces()
+        bc_type = np.asarray(bf.size * ["neu"])
+
+        xf = g.face_centers[:, bf]
+        xleft = np.where(xf[0] < 1e-3 + xf[0].min())[0]
+        xright = np.where(xf[0] > xf[0].max() - 1e-3)[0]
+        bc_type[xleft] = "dir"
+
+        bound = pp.BoundaryCondition(g, bf, bc_type)
+
+        bc_val = np.zeros(g.num_faces)
+        bc_val[bf[xright]] = -1
+
+        data = self.make_dictionary(g, bound, bc_val)
+        p = self.pressure(g, data)
+
+        bound_p = self.boundary_pressure(p, bc_val, data)
+        assert np.allclose(bound_p[bf], g.face_centers[0, bf])
+
+
+class TestMpfaPressureReconstructionMatrices:
+    def _make_true_2d(self, g):
+        if g.dim == 2:
+            g = g.copy()
+            g.cell_centers = np.delete(g.cell_centers, (2), axis=0)
+            g.face_centers = np.delete(g.face_centers, (2), axis=0)
+            g.face_normals = np.delete(g.face_normals, (2), axis=0)
+            g.nodes = np.delete(g.nodes, (2), axis=0)
+        return g
+
+    @property
+    def reference_dense_arrays(self):
+        return pp.test_utils.reference_dense_arrays.test_mpfa[
+            "TestMpfaPressureReconstructionMatrices"
+        ]
+
+    def test_cart_2d(self):
+        """
+        Test that mpfa gives out the correct matrices for reconstruction of the
+        pressures at the faces. Also check those returned by the helper function
+        reconstruct_pressure.
+        """
+        g = pp.CartGrid(np.array([1, 1]), physdims=[2, 2])
+        g.compute_geometry()
+
+        k = pp.SecondOrderTensor(np.array([2]))
+
+        bc = pp.BoundaryCondition(g)
+
+        _, _, grad_cell, grad_bound, *_ = discr_instance._flux_discretization(
+            g, k, bc, inverter="python"
+        )
+
+        reference_grad_bound = self.reference_dense_arrays["test_cart_2d"]["grad_bound"]
+        reference_grad_cell = np.array([[1.0], [1.0], [1.0], [1.0]])
+
+        assert np.all(np.abs(grad_bound - reference_grad_bound) < 1e-7)
+        assert np.all(np.abs(grad_cell - reference_grad_cell) < 1e-12)
+
+        # The reconstruction function requires a "true 2d" grid without z-coordinates.
+        g = self._make_true_2d(g)
+        sc_top = pp.fvutils.SubcellTopology(g)
+        D_g, CC = pp.numerics.fv.mpfa.reconstruct_presssure(g, sc_top, eta=0)
+
+        D_g_ref = self.reference_dense_arrays["test_cart_2d"]["D_g"]
+        CC_ref = np.array([[1.0], [1.0], [1.0], [1.0], [1.0], [1.0], [1.0], [1.0]])
+        assert np.all(np.abs(D_g - D_g_ref) < 1e-12)
+        assert np.all(np.abs(CC - CC_ref) < 1e-12)
+
+    def test_simplex_2d(self):
+        """Test reconstruct_pressure matrices"""
+        nx = 1
+        ny = 1
+        g = pp.StructuredTriangleGrid([nx, ny], physdims=[1, 1])
+        g.compute_geometry()
+        g = self._make_true_2d(g)
+        sc_top = pp.fvutils.SubcellTopology(g)
+
+        D_g, CC = pp.numerics.fv.mpfa.reconstruct_presssure(g, sc_top, eta=0)
+
+        references = self.reference_dense_arrays["test_simplex_2d"]
+        assert np.all(np.abs(D_g - references["D_g"]).A < 1e-12)
+        assert np.all(np.abs(CC - references["CC"]) < 1e-12)
+
+    def test_cart_3d(self):
+        """Test reconstruct_pressure matrices"""
+
+        g = pp.CartGrid([1, 1, 1], physdims=[2, 2, 2])
+        g.compute_geometry()
+        sc_top = pp.fvutils.SubcellTopology(g)
+
+        D_g, CC = pp.numerics.fv.mpfa.reconstruct_presssure(g, sc_top, eta=1)
+        references = self.reference_dense_arrays["test_cart_3d"]
+        assert np.all(np.abs(D_g - references["D_g"]).A < 1e-12)
+        assert np.all(np.abs(CC - references["CC"]) < 1e-12)
+
+    def test_simplex_3d_dirichlet(self):
+        """
+        Test that we retrieve a linear solution exactly
+        """
+        num_cells = 2 * np.ones(3, dtype=int)
+        g = pp.StructuredTetrahedralGrid(num_cells, physdims=[1, 1, 1])
+        g.compute_geometry()
+
+        kxx = np.ones(g.num_cells)
+        k = pp.SecondOrderTensor(kxx)
+
+        bc = pp.BoundaryCondition(g)
+        bc.is_dir[g.get_all_boundary_faces()] = True
+        bc.is_neu[bc.is_dir] = False
+
+        p0 = 1
+        p_b = np.sum(g.face_centers, axis=0) + p0
+
+        flux, bound_flux, p_t_cell, p_t_bound, *_ = discr_instance._flux_discretization(
+            g, k, bc, eta=0, inverter="python"
+        )
+
+        div = pp.fvutils.scalar_divergence(g)
+
+        P = sps.linalg.spsolve(div * flux, -div * bound_flux * p_b)
+
+        P_f = p_t_cell * P + p_t_bound * p_b
+
+        assert np.all(np.abs(P - np.sum(g.cell_centers, axis=0) - p0) < 1e-10)
+        assert np.all(np.abs(P_f - np.sum(g.face_centers, axis=0) - p0) < 1e-10)
+
+    def test_simplex_3d_boundary(self):
+        """
+        Even if we do not get exact solution at interiour we should be able to
+        retrieve the boundary conditions
+        """
+        num_cells = 2 * np.ones(3, dtype=int)
+        g = pp.StructuredTetrahedralGrid(num_cells, physdims=[1, 1, 1])
+        g.compute_geometry()
+
+        np.random.seed(2)
+
+        kxx = 10 * np.random.rand(g.num_cells)
+        k = pp.SecondOrderTensor(kxx)
+
+        bc = pp.BoundaryCondition(g)
+        dir_ind = g.get_all_boundary_faces()[[0, 2, 5, 8, 10, 13, 15, 21]]
+
+        bc.is_dir[dir_ind] = True
+        bc.is_neu[bc.is_dir] = False
+
+        p_b = np.random.randn(g.face_centers.shape[1])
+
+        flux, bound_flux, p_t_cell, p_t_bound, *_ = discr_instance._flux_discretization(
+            g, k, bc, eta=0, inverter="python"
+        )
+
+        div = pp.fvutils.scalar_divergence(g)
+
+        P = sps.linalg.spsolve(div * flux, -div * bound_flux * p_b.ravel("F"))
+
+        P_f = p_t_cell * P + p_t_bound * p_b
+
+        assert np.all(np.abs(P_f[dir_ind] - p_b[dir_ind]) < 1e-10)
+
+    def test_simplex_3d_sub_face(self):
+        """
+        Test that we reconstruct the exact solution on subfaces
+        """
+        num_cells = 2 * np.ones(3, dtype=int)
+        g = pp.StructuredTetrahedralGrid(num_cells, physdims=[1, 1, 1])
+        g.compute_geometry()
+        s_t = pp.fvutils.SubcellTopology(g)
+
+        k = pp.SecondOrderTensor(10 * np.ones(g.num_cells))
+
+        bc = pp.BoundaryCondition(g)
+        bc.is_dir[g.get_all_boundary_faces()] = True
+        bc.is_neu[bc.is_dir] = False
+        bc = pp.fvutils.boundary_to_sub_boundary(bc, s_t)
+
+        p_b = np.sum(-g.face_centers, axis=0)
+        p_b = p_b[s_t.fno_unique]
+
+        flux, bound_flux, p_t_cell, p_t_bound, *_ = discr_instance._flux_discretization(
+            g, k, bc, eta=0, inverter="python"
+        )
+
+        div = pp.fvutils.scalar_divergence(g)
+
+        hf2f = pp.fvutils.map_hf_2_f(nd=1, sd=g)
+        P = sps.linalg.spsolve(div * hf2f * flux, -div * hf2f * bound_flux * p_b)
+
+        P_hf = p_t_cell * P + p_t_bound * p_b
+        _, IA = np.unique(s_t.fno_unique, True)
+        P_f = P_hf[IA]
+
+        assert np.all(np.abs(P + np.sum(g.cell_centers, axis=0)) < 1e-10)
+        assert np.all(np.abs(P_f + np.sum(g.face_centers, axis=0)) < 1e-10)
